refactor: 스폰서 섹션 컴포넌트 분리
